Tag search refinements
Added quite a few more test cases for tag search, including rails and
angular searches. Modified the search algorithm to hopefully be even
more accurate for how we want our search results working. Slugs with a
‘-‘ in them are now not considered “starting with”. For example, if you
searched for “iOS” and there was a slug in there for “ios-simulator” it
would not be ranked as starting with “iOS” and thus would give the user
what I believe are more accurate results: “iOS”, “ios-4”, and “ios-5”.

However, when searching for “iOS simulator” it comes to the top.