Modified exact match weight
Modified the weight for an exact match to be heavier than anything
else. I further added the “ruby” tag to the test list to ensure that
the automated ruby test would succeed. Finally I removed the second the
third place matches because that’s not consistent across dev boxes
because if the dev box has more data than the test data it won’t be
consistent; the important thing is the top match.